app/pyproject.toml
[tool.poetry]
name = "libs"
version = "0.1"
description = "poetry libs for the project"
authors = ["Ilia Britov <ilia.brittov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.9"
requests = "^2.22.0"
flask = "^1.1.1"
sqlalchemy = "1.3.8"
psycopg2-binary = "^2.8.4"
pytest = "^5.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
